feat: protocol update for 13.20 (#1269)
Upgrade the client protocol to 13.20 with the necessary protocol changes, including the variable type change for magical effects from uint8_t to uint16_t, since on this new version new effects were added, and they reached the 'byte' limit (255).

Client download link:
https://github.com/dudantas/tibia-client/releases/tag/13.20.13560

NOTE from client:
• Use the client "client-127.0.0.1" to test locally, it already uses the Open Tibia RSA key and the local ip.
• Place the files "cache/eventscheduler.json" and "cache/boostedcreature.json" as read-only for the correct functioning of the systems by the server.
• Download client 13.20 from "Source code (zip)".
• Download client 10x from "client-11-outdated-(10x).zip"

+++ src/creatures/combat/combat.cpp
@@ -414,12 +414,12 @@ bool Combat::setParam(CombatParam_t param, uint32_t value) {
 		}
 
 		case COMBAT_PARAM_EFFECT: {
-			params.impactEffect = static_cast<uint8_t>(value);
+			params.impactEffect = static_cast<uint16_t>(value);
 			return true;
 		}
 
 		case COMBAT_PARAM_DISTANCEEFFECT: {
-			params.distanceEffect = static_cast<uint8_t>(value);
+			params.distanceEffect = static_cast<uint16_t>(value);
 			return true;
 		}
 
@@ -851,7 +851,7 @@ void Combat::postCombatEffects(Creature* caster, const Position &origin, const P
 	}
 }
 
-void Combat::addDistanceEffect(Creature* caster, const Position &fromPos, const Position &toPos, uint8_t effect) {
+void Combat::addDistanceEffect(Creature* caster, const Position &fromPos, const Position &toPos, uint16_t effect) {
 	if (effect == CONST_ANI_WEAPONTYPE) {
 		if (!caster) {
 			return;
